Rename {run.rs => bot.rs}
